sentiment_eval/requirements.txt
python-dotenv>=1.0.0
pyyaml>=6.0
pandas>=2.0.0
requests>=2.31.0
tqdm>=4.65.0
streamlit>=1.30.0
matplotlib>=3.8.0
scikit-learn>=1.3.0